--- crates/miden-standards/asm/account_components/faucets/basic_fungible_faucet.masm
@@ -1,25 +1,0 @@
-# The MASM code of the Basic Fungible Faucet Account Component.
-#
-# See the `BasicFungibleFaucet` Rust type's documentation for more details. This component
-# bundles the fungible-faucet asset minting/burning procedures and the token metadata accessors
-# (name, description, logo URI, external link, mutability config) into a single component.
-
-pub use ::miden::standards::faucets::basic_fungible::mint_and_send
-pub use ::miden::standards::faucets::basic_fungible::receive_and_burn
-
-# Token config + name + optional fields accessors and owner-gated setters.
-pub use ::miden::standards::faucets::fungible::get_name
-pub use ::miden::standards::faucets::fungible::get_mutability_config
-pub use ::miden::standards::faucets::fungible::is_max_supply_mutable
-pub use ::miden::standards::faucets::fungible::is_description_mutable
-pub use ::miden::standards::faucets::fungible::is_logo_uri_mutable
-pub use ::miden::standards::faucets::fungible::is_external_link_mutable
-pub use ::miden::standards::faucets::fungible::get_token_metadata
-pub use ::miden::standards::faucets::fungible::get_max_supply
-pub use ::miden::standards::faucets::fungible::get_decimals
-pub use ::miden::standards::faucets::fungible::get_token_symbol
-pub use ::miden::standards::faucets::fungible::get_token_supply
-pub use ::miden::standards::faucets::fungible::set_description
-pub use ::miden::standards::faucets::fungible::set_logo_uri
-pub use ::miden::standards::faucets::fungible::set_external_link
-pub use ::miden::standards::faucets::fungible::set_max_supply
