--- crates/miden-standards/asm/standards/faucets/basic_fungible.masm
@@ -1,64 +1,0 @@
-# BASIC FUNGIBLE FAUCET CONTRACT
-#
-# See the `BasicFungibleFaucet` documentation for details.
-#
-# Note: This component requires mint-policy and burn-policy components to also be present in the
-# account.
-# =================================================================================================
-
-use miden::standards::faucets
-use miden::standards::faucets::policies::policy_manager
-
-# PROCEDURES
-# =================================================================================================
-
-#! Mints fungible assets to the provided recipient by creating a note.
-#!
-#! This procedure first executes the active mint policy configured via the `TokenPolicyManager`
-#! and then mints the asset and creates an output note with that asset for the recipient.
-#!
-#! Inputs:  [amount, tag, note_type, RECIPIENT, pad(9)]
-#! Outputs: [note_idx, pad(15)]
-#!
-#! Where:
-#! - amount is the amount to be minted and sent.
-#! - tag is the tag to be included in the note.
-#! - note_type is the type of the note that holds the asset.
-#! - RECIPIENT is the recipient of the asset, i.e.,
-#!   hash(hash(hash(serial_num, [0; 4]), script_root), storage_commitment).
-#! - note_idx is the index of the created note.
-#!
-#! Panics if:
-#! - active mint policy validation fails.
-#! - any of the validations in faucets::mint_and_send fail.
-#!
-#! Invocation: call
-pub proc mint_and_send
-    exec.policy_manager::execute_mint_policy
-    # => [new_amount, new_tag, new_note_type, NEW_RECIPIENT, pad(9)]
-
-    exec.faucets::mint_and_send
-    # => [note_idx, pad(15)]
-end
-
-#! Receives a fungible asset from the active note and burns it.
-#!
-#! This is a re-export of `miden::standards::faucets::receive_and_burn`.
-#!
-#! The shared procedure first executes the active burn policy and then burns the single
-#! fungible asset contained in the active note.
-#!
-#! Inputs:  [pad(16)]
-#! Outputs: [pad(16)]
-#!
-#! Panics if:
-#! - active burn policy validation fails.
-#! - the procedure is not called from a note context (active_note::get_assets will fail).
-#! - the note does not contain exactly one asset.
-#! - the transaction is executed against an account which is not a fungible asset faucet.
-#! - the transaction is executed against a faucet which is not the origin of the specified asset.
-#! - the amount about to be burned is greater than the outstanding supply of the asset.
-#! - any of the validations in faucets::receive_and_burn fail.
-#!
-#! Invocation: call
-pub use faucets::receive_and_burn
