+++ crates/miden-standards/asm/standards/notes/mint.masm
@@ -1,6 +1,6 @@
 use miden::protocol::active_note
 use miden::protocol::note
-use miden::standards::faucets::basic_fungible->faucet
+use miden::standards::faucets::fungible->faucet
 
 # CONSTANTS
 # =================================================================================================
@@ -34,7 +34,7 @@ const ERR_MINT_UNEXPECTED_NUMBER_OF_STORAGE_ITEMS="MINT script expects exactly 6
 #! This note is intended to be executed against a network fungible faucet account.
 #!
 #! Requires that the account exposes:
-#! - miden::standards::faucets::basic_fungible::mint_and_send procedure.
+#! - miden::standards::faucets::fungible::mint_and_send procedure.
 #!
 #! Inputs:  [ARGS, pad(12)]
 #! Outputs: [pad(16)]
