+++ crates/miden-standards/src/account/metadata/mod.rs
@@ -1,5 +1,3 @@
 mod schema_commitment;
-mod token_metadata;
 
 pub use schema_commitment::{AccountBuilderSchemaCommitmentExt, AccountSchemaCommitment};
-pub use token_metadata::{Description, ExternalLink, LogoURI, TokenMetadata, TokenName};
