+++ crates/miden-testing/src/kernel_tests/tx/test_callbacks.rs
@@ -31,8 +31,7 @@ use miden_protocol::errors::MasmError;
 use miden_protocol::note::{NoteTag, NoteType};
 use miden_protocol::utils::sync::LazyLock;
 use miden_protocol::{Felt, Word};
-use miden_standards::account::faucets::BasicFungibleFaucet;
-use miden_standards::account::metadata::TokenName;
+use miden_standards::account::faucets::{FungibleFaucet, TokenName};
 use miden_standards::account::policies::{
     BurnPolicyConfig,
     MintPolicyConfig,
@@ -664,7 +663,7 @@ async fn test_faucet_with_callback_calls_itself() -> anyhow::Result<()> {
             push.{amount}
             # => [amount, tag, note_type, RECIPIENT, pad(9)]
 
-            call.::miden::standards::faucets::basic_fungible::mint_and_send
+            call.::miden::standards::faucets::fungible::mint_and_send
             # => [note_idx, pad(15)]
 
             # truncate the stack
@@ -759,11 +758,11 @@ fn add_faucet_with_callbacks(
         callbacks = callbacks.on_before_asset_added_to_note(proc_root);
     }
 
-    let faucet = BasicFungibleFaucet::builder(
+    let faucet = FungibleFaucet::builder(
         TokenName::new("").expect("empty string is a valid token name"),
         "SYM".try_into()?,
         8,
-        1_000_000u64,
+        miden_protocol::asset::AssetAmount::new(1_000_000)?,
     )
     .build()?;
