+++ crates/miden-testing/tests/scripts/faucet.rs
@@ -30,8 +30,7 @@ use miden_protocol::testing::account_id::ACCOUNT_ID_PRIVATE_SENDER;
 use miden_protocol::transaction::{ExecutedTransaction, RawOutputNote};
 use miden_protocol::{Felt, Word};
 use miden_standards::account::access::Ownable2Step;
-use miden_standards::account::faucets::BasicFungibleFaucet;
-use miden_standards::account::metadata::TokenName;
+use miden_standards::account::faucets::{FungibleFaucet, TokenName};
 use miden_standards::account::policies::{
     BurnAllowAll,
     BurnOwnerOnly,
@@ -87,7 +86,7 @@ pub fn create_mint_script_code(params: &FaucetTestParams) -> String {
                 push.{amount}
                 # => [amount, tag, note_type, RECIPIENT, pad(9)]
 
-                call.::miden::standards::faucets::basic_fungible::mint_and_send
+                call.::miden::standards::faucets::fungible::mint_and_send
                 # => [note_idx, pad(15)]
 
                 # truncate the stack
@@ -197,7 +196,9 @@ fn build_network_faucet_with_burn_switching(
 ) -> anyhow::Result<Account> {
     let name = TokenName::new(token_symbol)?;
     let symbol = TokenSymbol::new(token_symbol)?;
-    let metadata = BasicFungibleFaucet::builder(name, symbol, 10, max_supply)
+    let max_supply = miden_protocol::asset::AssetAmount::new(max_supply)?;
+    let token_supply = miden_protocol::asset::AssetAmount::new(token_supply)?;
+    let metadata = FungibleFaucet::builder(name, symbol, 10, max_supply)
         .token_supply(token_supply)
         .build()?;
 
@@ -282,7 +283,7 @@ async fn faucet_contract_mint_fungible_asset_fails_exceeds_max_supply() -> anyho
                 push.{amount}
                 # => [amount, tag, note_type, RECIPIENT, pad(9)]
 
-                call.::miden::standards::faucets::basic_fungible::mint_and_send
+                call.::miden::standards::faucets::fungible::mint_and_send
                 # => [note_idx, pad(15)]
 
                 # truncate the stack
@@ -365,7 +366,7 @@ async fn prove_burning_fungible_asset_on_existing_faucet_succeeds() -> anyhow::R
             dropw
             # => []
 
-            call.::miden::standards::faucets::basic_fungible::receive_and_burn
+            call.::miden::standards::faucets::fungible::receive_and_burn
             # => [pad(16)]
         end
         ";
@@ -375,7 +376,7 @@ async fn prove_burning_fungible_asset_on_existing_faucet_succeeds() -> anyhow::R
     builder.add_output_note(RawOutputNote::Full(note.clone()));
     let mock_chain = builder.build()?;
 
-    let token_metadata = BasicFungibleFaucet::try_from(faucet.storage())?;
+    let token_metadata = FungibleFaucet::try_from(faucet.storage())?;
 
     // Check that max_supply at the word's index 0 is 200. The remainder of the word is initialized
     // with the metadata of the faucet which we don't need to check.
@@ -429,7 +430,7 @@ async fn faucet_burn_fungible_asset_fails_amount_exceeds_token_supply() -> anyho
             dropw
             # => []
 
-            call.::miden::standards::faucets::basic_fungible::receive_and_burn
+            call.::miden::standards::faucets::fungible::receive_and_burn
             # => [pad(16)]
         end
         ";
@@ -540,7 +541,7 @@ async fn test_public_note_creation_with_script_from_datastore() -> anyhow::Resul
                 push.{amount}
                 # => [amount, tag, note_type, RECIPIENT]
 
-                call.::miden::standards::faucets::basic_fungible::mint_and_send
+                call.::miden::standards::faucets::fungible::mint_and_send
                 # => [note_idx, pad(15)]
 
                 # Truncate the stack
@@ -656,7 +657,7 @@ async fn network_faucet_mint() -> anyhow::Result<()> {
     let mut target_account = builder.add_existing_wallet(Auth::IncrNonce)?;
 
     // Check the Network Fungible Faucet's max supply.
-    let actual_max_supply = BasicFungibleFaucet::try_from(faucet.storage())?.max_supply();
+    let actual_max_supply = FungibleFaucet::try_from(faucet.storage())?.max_supply();
     assert_eq!(actual_max_supply.as_canonical_u64(), max_supply);
 
     // Check that the creator account ID is stored in the ownership slot.
@@ -672,7 +673,7 @@ async fn network_faucet_mint() -> anyhow::Result<()> {
 
     // Check that the faucet's token supply has been correctly initialized.
     // The already issued amount should be 50.
-    let initial_token_supply = BasicFungibleFaucet::try_from(faucet.storage())?.token_supply();
+    let initial_token_supply = FungibleFaucet::try_from(faucet.storage())?.token_supply();
     assert_eq!(initial_token_supply.as_canonical_u64(), token_supply);
 
     // CREATE MINT NOTE USING STANDARD NOTE
@@ -828,7 +829,7 @@ async fn test_network_faucet_set_policy_rejects_non_allowed_root() -> anyhow::Re
     let mock_chain = builder.build()?;
 
     // This root exists in account code, but is not in the mint policy allowlist.
-    let invalid_policy_root = BasicFungibleFaucet::mint_and_send_digest();
+    let invalid_policy_root = FungibleFaucet::mint_and_send_digest();
     let set_policy_note_script = format!(
         r#"
         use miden::standards::faucets::policies::policy_manager
@@ -879,7 +880,7 @@ async fn test_network_faucet_set_burn_policy_rejects_non_allowed_root() -> anyho
     let mock_chain = builder.build()?;
 
     // This root exists in account code, but is not in the burn policy allowlist.
-    let invalid_policy_root = BasicFungibleFaucet::receive_and_burn_digest();
+    let invalid_policy_root = FungibleFaucet::receive_and_burn_digest();
     let set_policy_note_script = create_set_burn_policy_note_script(invalid_policy_root);
 
     let result = execute_faucet_note_script(
@@ -1348,8 +1349,8 @@ fn test_faucet_burn_procedures_are_identical() {
     // Both faucet types must export the same burn procedure with identical MAST roots
     // so that a single BURN note script can work with either faucet type
     assert_eq!(
-        BasicFungibleFaucet::receive_and_burn_digest(),
-        BasicFungibleFaucet::receive_and_burn_digest(),
+        FungibleFaucet::receive_and_burn_digest(),
+        FungibleFaucet::receive_and_burn_digest(),
         "Basic and network fungible faucets must have the same burn procedure digest"
     );
 }
@@ -1421,7 +1422,7 @@ async fn network_faucet_burn() -> anyhow::Result<()> {
     mock_chain.prove_next_block()?;
 
     // Check the initial token issuance before burning
-    let initial_token_supply = BasicFungibleFaucet::try_from(faucet.storage())?.token_supply();
+    let initial_token_supply = FungibleFaucet::try_from(faucet.storage())?.token_supply();
     assert_eq!(initial_token_supply, Felt::new(100));
 
     // EXECUTE BURN NOTE AGAINST NETWORK FAUCET
@@ -1438,7 +1439,7 @@ async fn network_faucet_burn() -> anyhow::Result<()> {
 
     // Apply the delta to the faucet account and verify the token issuance decreased
     faucet.apply_delta(executed_transaction.account_delta())?;
-    let final_token_supply = BasicFungibleFaucet::try_from(faucet.storage())?.token_supply();
+    let final_token_supply = FungibleFaucet::try_from(faucet.storage())?.token_supply();
     assert_eq!(
         final_token_supply,
         Felt::new(initial_token_supply.as_canonical_u64() - burn_amount)
@@ -1722,7 +1723,7 @@ async fn multiple_mints_in_single_tx_produce_correct_amounts() -> anyhow::Result
                 push.{amount_1}
                 # => [amount_1, tag, note_type, RECIPIENT_1, pad(9)]
 
-                call.::miden::standards::faucets::basic_fungible::mint_and_send
+                call.::miden::standards::faucets::fungible::mint_and_send
                 # => [note_idx, pad(15)]
 
                 # clean up the stack before the second call
@@ -1737,7 +1738,7 @@ async fn multiple_mints_in_single_tx_produce_correct_amounts() -> anyhow::Result
                 push.{amount_2}
                 # => [amount_2, tag, note_type, RECIPIENT_2, pad(9)]
 
-                call.::miden::standards::faucets::basic_fungible::mint_and_send
+                call.::miden::standards::faucets::fungible::mint_and_send
                 # => [note_idx, pad(15)]
 
                 # truncate the stack
